Extracted Thread to its own file
